+++ tests/test_potentials.py
@@ -576,22 +576,22 @@ def test_combined_potential_learnable_weights():
 
 
 @pytest.mark.parametrize("device", ["cpu", "cuda"])
+@pytest.mark.parametrize("dtype", [torch.float32, torch.float64])
 @pytest.mark.parametrize(
     "potential_class", [CoulombPotential, InversePowerLawPotential, SplinePotential]
 )
-def test_potential_device(potential_class, device):
+def test_potential_device_dtype(potential_class, device, dtype):
     if device == "cuda" and not torch.cuda.is_available():
         pytest.skip("CUDA is not available")
 
     smearing = 1.0
     exponent = 1.0
-    dtype = torch.float64
 
     if potential_class is InversePowerLawPotential:
         potential = potential_class(
             exponent=exponent, smearing=smearing, dtype=dtype, device=device
         )
-    elif potential_class == SplinePotential:
+    elif potential_class is SplinePotential:
         x_grid = torch.linspace(0, 20, 100, device=device, dtype=dtype)
         y_grid = torch.exp(-(x_grid**2) * 0.5)
         potential = potential_class(
@@ -604,3 +604,4 @@ def test_potential_device(potential_class, device):
     potential_lr = potential.lr_from_dist(dists)
 
     assert potential_lr.device.type == device
+    assert potential_lr.dtype == dtype